moved search components into comp folder
